Print comments alongside "held back" gems
If a Gemfile has an inline or prefix comment for a particular gem, and
that gem is being held back due to a pin, then print the comment along
with the pin information in bundleup's output. Assuming the comment is
meaningful, this helps users understand *why* the gem is being held
back.