Install missing gems before attempting to update
The first thing that bundleup does to upgrade a Gemfile is to run the
`bundle list` command. This command fails if there are any dependencies
listed in the Gemfile that are not currently installed, causing bundleup
to crash.

This is annoying and happens fairly often if you are switching to a git
branch where another person (or dependabot) has recently upgraded or
added a dependency.

This commit fixes this crash by always running `bundle check` first,
before attempting to run `bundle list`. If the check fails, bundleup
will automatically run `bundle install` to ensure all dependencies are
installed before it performs an upgrade.