Remove implicit shell out to git in gemspec (#82)
Our gemspec file might be used in an environment like a Docker container
where `git` is not installed. Replace `git ls-files` external shell
command with Ruby's built-in `Dir.glob`.

`git ls-files` is still useful in preventing us from packaging files in
the gem inadvertently, so as a safety measure use it during the build
step as an enhancement to `rake release`.

Also disable overcommit during `rake release` so that we don't get
failures if `rake release` is run via `bundle exec`.